feature: add WordDatabase Protocol
