+++ JWords/DataService/WordService.swift
@@ -11,14 +11,7 @@ typealias FireStoreCompletion = ((Error?) -> Void)?
 
 class WordService {
     static func getWordBooks(completionHandler: @escaping ([WordBook]?, Error?) -> Void) {
-        Constants.Collections.wordBooks.order(by: "timestamp", descending: true).getDocuments { snapshot, error in
-            if let error = error {
-                completionHandler(nil, error)
-            }
-            guard let documents = snapshot?.documents else { return }
-            let wordBooks = documents.compactMap({ try? $0.data(as: WordBook.self) })
-            completionHandler(wordBooks, nil)
-        }
+        Firestore.getWordBooks(completionHandler: completionHandler)
     }
     
     static func getWords(wordBookID id: String, completionHandler: @escaping ([Word]?, Error?) -> Void) {
